fixing subscriptions namespace to avoid conflicts with model of the same name
