+++ src/Stripe.Client.Sdk.Tests/Clients/Subscriptions/PlanClientTests.cs
@@ -4,7 +4,7 @@ using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using Stripe.Client.Sdk.Clients;
-using Stripe.Client.Sdk.Clients.Subscription;
+using Stripe.Client.Sdk.Clients.Subscriptions;
 using Stripe.Client.Sdk.Models;
 using Stripe.Client.Sdk.Models.Arguments;
 using Stripe.Client.Sdk.Models.Filters;
