+++ src/Stripe.Client.Sdk/Clients/Subscriptions/PlanClient.cs
@@ -6,66 +6,66 @@ using Stripe.Client.Sdk.Models;
 using Stripe.Client.Sdk.Models.Arguments;
 using Stripe.Client.Sdk.Models.Filters;
 
-namespace Stripe.Client.Sdk.Clients.Subscription
+namespace Stripe.Client.Sdk.Clients.Subscriptions
 {
-    public class CouponClient : ICouponClient
+    public class PlanClient : IPlanClient
     {
         private readonly IStripeClient _client;
 
-        public CouponClient(IStripeClient client)
+        public PlanClient(IStripeClient client)
         {
             _client = client;
         }
 
-        public async Task<StripeResponse<Coupon>> GetCoupon(string couponId,
+        public async Task<StripeResponse<Plan>> GetPlan(string planId,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var request = new StripeRequest<Coupon>
+            var request = new StripeRequest<Plan>
             {
-                UrlPath = PathHelper.GetPath(Paths.Coupons, couponId)
+                UrlPath = PathHelper.GetPath(Paths.Plans, planId)
             };
             return await _client.Get(request, cancellationToken);
         }
 
-        public async Task<StripeResponse<Pagination<Coupon>>> GetCoupons(CouponListFilter filter,
+        public async Task<StripeResponse<Pagination<Plan>>> GetPlans(PlanListFilter filter,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var request = new StripeRequest<CouponListFilter, Pagination<Coupon>>
+            var request = new StripeRequest<PlanListFilter, Pagination<Plan>>
             {
-                UrlPath = Paths.Coupons,
+                UrlPath = Paths.Plans,
                 Model = filter
             };
             return await _client.Get(request, cancellationToken);
         }
 
-        public async Task<StripeResponse<Coupon>> CreateCoupon(CouponCreateArguments arguments,
+        public async Task<StripeResponse<Plan>> CreatePlan(PlanCreateArguments arguments,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var request = new StripeRequest<CouponCreateArguments, Coupon>
+            var request = new StripeRequest<PlanCreateArguments, Plan>
             {
-                UrlPath = Paths.Coupons,
+                UrlPath = Paths.Plans,
                 Model = arguments
             };
             return await _client.Post(request, cancellationToken);
         }
 
-        public async Task<StripeResponse<Coupon>> UpdateCoupon(CouponUpdateArguments arguments,
+        public async Task<StripeResponse<Plan>> UpdatePlan(PlanUpdateArguments arguments,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var request = new StripeRequest<CouponUpdateArguments, Coupon>
+            var request = new StripeRequest<PlanUpdateArguments, Plan>
             {
-                UrlPath = PathHelper.GetPath(Paths.Coupons, arguments.CouponId),
+                UrlPath = PathHelper.GetPath(Paths.Plans, arguments.PlanId),
                 Model = arguments
             };
             return await _client.Post(request, cancellationToken);
         }
 
-        public async Task<StripeResponse<DeletedObject>> DeleteCoupon(string couponId,
+        public async Task<StripeResponse<DeletedObject>> DeletePlan(string planId,
             CancellationToken cancellationToken = default(CancellationToken))
         {
             var request = new StripeRequest<DeletedObject>
             {
-                UrlPath = PathHelper.GetPath(Paths.Coupons, couponId)
+                UrlPath = PathHelper.GetPath(Paths.Plans, planId)
             };
             return await _client.Delete(request, cancellationToken);
         }
