+++ src/Stripe.Client.Sdk/Stripe.Client.Sdk.csproj
@@ -55,8 +55,8 @@
     <Compile Include="Clients\Connect\AccountClient.cs" />
     <Compile Include="Clients\Core\IRefundClient.cs" />
     <Compile Include="Clients\Core\RefundClient.cs" />
-    <Compile Include="Clients\Subscription\IInvoiceItemClient.cs" />
-    <Compile Include="Clients\Subscription\InvoiceItemClient.cs" />
+    <Compile Include="Clients\Subscriptions\IInvoiceItemClient.cs" />
+    <Compile Include="Clients\Subscriptions\InvoiceItemClient.cs" />
     <Compile Include="Configuration\AppSettingsConfiguration.cs" />
     <Compile Include="Configuration\IStripeConfiguration.cs" />
     <Compile Include="Configuration\StripeConfiguration.cs" />
@@ -153,7 +153,7 @@
     <Compile Include="Clients\Core\BalanceClient.cs" />
     <Compile Include="Clients\Payment\BitcoinReceiverClient.cs" />
     <Compile Include="Clients\Core\ChargeClient.cs" />
-    <Compile Include="Clients\Subscription\CouponClient.cs" />
+    <Compile Include="Clients\Subscriptions\CouponClient.cs" />
     <Compile Include="Clients\Core\CustomerClient.cs" />
     <Compile Include="Clients\Core\DisputeClient.cs" />
     <Compile Include="Clients\Core\EventClient.cs" />
@@ -163,15 +163,15 @@
     <Compile Include="Clients\Core\IBalanceClient.cs" />
     <Compile Include="Clients\Payment\IBitcoinReceiverClient.cs" />
     <Compile Include="Clients\Core\IChargeClient.cs" />
-    <Compile Include="Clients\Subscription\ICouponClient.cs" />
+    <Compile Include="Clients\Subscriptions\ICouponClient.cs" />
     <Compile Include="Clients\Core\ICustomerClient.cs" />
     <Compile Include="Clients\Core\IDisputeClient.cs" />
     <Compile Include="Clients\Core\IEventClient.cs" />
     <Compile Include="Clients\Core\IFileUploadClient.cs" />
-    <Compile Include="Clients\Subscription\IInvoiceClient.cs" />
-    <Compile Include="Clients\Subscription\InvoiceClient.cs" />
+    <Compile Include="Clients\Subscriptions\IInvoiceClient.cs" />
+    <Compile Include="Clients\Subscriptions\InvoiceClient.cs" />
     <Compile Include="Clients\Relay\IOrderClient.cs" />
-    <Compile Include="Clients\Subscription\IPlanClient.cs" />
+    <Compile Include="Clients\Subscriptions\IPlanClient.cs" />
     <Compile Include="Clients\Relay\IProductClient.cs" />
     <Compile Include="Clients\Connect\IRecipientClient.cs" />
     <Compile Include="Clients\Relay\ISkuClient.cs" />
@@ -179,7 +179,7 @@
     <Compile Include="Clients\Core\ITokenClient.cs" />
     <Compile Include="Clients\Core\ITransferClient.cs" />
     <Compile Include="Clients\Relay\OrderClient.cs" />
-    <Compile Include="Clients\Subscription\PlanClient.cs" />
+    <Compile Include="Clients\Subscriptions\PlanClient.cs" />
     <Compile Include="Clients\Relay\ProductClient.cs" />
     <Compile Include="Clients\Connect\RecipientClient.cs" />
     <Compile Include="Clients\Relay\SkuClient.cs" />
